refactor: split provider into provider and provider/protocol
