+++ internal/provider/protocol/base.go
@@ -1,0 +1,3 @@
+// Package protocol implements all the generic protocols to detect public IP addresses.
+// Generic protocols are not tied to a particular server (e.g., Cloudflare).
+package protocol
